requirements.txt
cycler==0.11.0
kiwisolver==1.4.4
numpy==1.25.1
opencv-python==4.8.0.74
Pillow==10.0.0
pyparsing==3.1.0
pytesseract==0.3.7
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
